requirements.txt
fastapi
uvicorn
duckdb
python-dotenv
